ref(ui): Move components to more logical folders
Also codesplits the ProjectEnvironments view
