+++ src/sentry/static/sentry/app/components/avatar/actorAvatar.jsx
