+++ src/sentry/static/sentry/app/components/issues/compactIssue.jsx
