+++ src/sentry/static/sentry/app/components/projects/bookmarkStar.jsx
